Rename input.rs to button_input.rs
